NOTE: This port belongs in textproc/
Add libcroco. I'm starting it off in this repository because I
haven't the slightest clue what its use is.

As a CSS2 library, it seems like it wants to be accessible to
a wide range of applications. Thing is, it depends upon pango.
For now, I've put it into ${LOCALBASE}, but it might need to
go into ${X11BASE}.


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@971 df743ca5-7f9a-e211-a948-0013205c9059