Re-add bsd.gnome.mk.
This has added 4 new macro targets: pygtk, pygnome, pygtk2, and pygnome2.

They work fine, but you have to define USE_PYTHON in the port Makefile.
I'm not sure whether that is good or bad. On the one hand, having it
happen automatically makes it easier to port. But on the other hand,
it's nice to be able to recursively grep for USE_PYTHON and find all the
python-related ports that way.


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@1230 df743ca5-7f9a-e211-a948-0013205c9059