Check to see whether we're upgrading from an older version of glib (i.e., not glib-2.3). If so, and FORCE_MANUAL_UPGRADE is not defined, error out and instruct people to use gnome_upgrade.sh.
Right now it tells people to read develfaq.html, but it should be changed
to the hand-holding help document that I'm going to finish working on
one of these days/weeks/months.

This calls pkg_info|grep, so it can make it seem like stuff is dragging
at first. Perhaps there's a better way to run this call?


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@1939 df743ca5-7f9a-e211-a948-0013205c9059