Add evince-0.1.0.
This has been build-tested only, as I can't boot into FreeBSD at the
moment. Note also that I'm not even sure exactly what this application
does, because the authors didn't see fit to include a single descriptive
line anywhere, nor create a webpage, etc.

I'm also not sure if the category is appropriate.

Please treat this port as a framework, and feel free to edit/update/remove
it as you see fit ::P


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@3366 df743ca5-7f9a-e211-a948-0013205c9059