Add pkg-config at 0.18.1.
I'm adding it here rather than into the FreeBSD CVS repo because during
the last pkg-config update, there were some breakages. Also, there is
some handling for potentially different --libs output, so it's best that
it percolate here for a while. It's probably safe to wait and merge along
with the gnome-2.12 update.


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@4192 df743ca5-7f9a-e211-a948-0013205c9059