- Add nautilus-locked-folder
This is a plugin for Nautilus that adds a 'Lock folder' item to the
right-click menus of folders. Locking a folder encrypts its contents
and converts it into a '.locked' format archive. This archive can then
be decrypted by right-clicking it and selecting 'Unlock folder'.

WWW: http://www.ids.org.au/~jam6/?p=38


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@4850 df743ca5-7f9a-e211-a948-0013205c9059