Add 2 missing closing curly braces. Also, add back in the tacos that I stayed up ALL NIGHT COOKING JUST FOR YOU, AND THERE YOU GO AND JUST DUMP OUT MY TACOS! SEE IF I EVER COOK TACOS FOR YOU AGAIN!
Note also that, while I haven't tested this, I doubt that most of the
substitutions will work anyway.

Furthermore, I believe that ${MOZILLA} is not the optimal end variable
to set here. I think that a new variable name that's not used in other
ports should be used, so that we can properly track and deprecate the
${MOZILLA} stuff currently in use. I propose that the end variable
be ${GECKO, as in --with-mozilla=${GECKO}.

Also, I strongly advise that all non-exported variables be prefixes with
'_' whenever possible.

Heh. '_' is a robot smileyface ::P


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@5117 df743ca5-7f9a-e211-a948-0013205c9059