Pessulus is a lockdown editor for GNOME, written in python. It enables administrators to set mandatory settings in GConf. The users can not change these settings. Use of pessulus can be useful on computers that are open to use by everyone, e.g. in an internet cafe.
WWW: http://www.gnome.org/~vuntz/pessulus/

--

It works great to disable/enable in both normal user and mandatory (admin),
to tweak the mandatory settings, of course you have to be login as root to
have the mandatory feature. If you run it as in normal user, the mandatory
feature will be hide (disable). Example, in the left is root that has
lock/unlock buttons in the left of check box and in the right is user that
has no lock/unlock button.

	http://people.freebsd.org/~mezz/pessulus.png

BTW: Going to try to work on sabayon next...


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@5950 df743ca5-7f9a-e211-a948-0013205c9059