Update to 2.0.2.
This is a kindof sketchy update. It's out of sync with glade3 and there hasn't
been much forward development from the authors. I've nuked the glade plugin
(patch-plugins_Makefile.in) and I've commented out the tag creation script
because it was doing something weird. The latter issue is probably more
easily fixed than the former.


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@7044 df743ca5-7f9a-e211-a948-0013205c9059