Update glib20 to 2.15.1, and break out the GIO FAM backend into a separate port (gio-fam-backend).
This requires a bit of trickery in bsd.gnome.mk to
avoid a dependency loop, so a new _glib20 component was added that should
only be required by FAM providers.  To that end, add gamin, and make gamin
depend on _glib20.

Update gvfs to 0.1.1, and add a new gvfs component to bsd.gnome.mk.

gio-standalone is no longer needed as this code is now part of glib.


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@10030 df743ca5-7f9a-e211-a948-0013205c9059