Okay, so here's the problem.  init runs the rc scripts before reading /etc/ttys.  This presents a problem to gdm and ConsoleKit because they need to access VT resources when they start.  When this doesn't happen, we have a problem.
To solve this, I created vty-checker.  Sounded like a good idea, except that
opening the ttyv* devices will always work -- even when the console isn't
ready.  So let's try a new approach.  Instead of looping on vty-checker,
loop on ps, and look for getty processes.  Do not start gdm and hald
until we see them.

If anyone has a better idea of something to check that will definitely tell
us when the console is fully initialized, PLEASE let me know.


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@10346 df743ca5-7f9a-e211-a948-0013205c9059