- Rename from webkit to webkit-gtk2. - Update to 30028, but I added '0.0.' as in 0.0.30028 that way we don't have   to use the PORTEPOCH in future. Also, this version makes a lot sense. - Complete seperate from webkit-qt4, the build process is different now in   this version by using autotools for gtk2 and qmake for qt4. - Remove 'gnome' from CATEGORIES. This is not a GNOME stuff. ;-)
I am keeping webkit for now to not break webkit-qt4. Also, Epiphany does not
build with this new version, which it's fixed in SVN. In next version of
Epiphany (if you select option) will have to use webkit-gtk2 to be able to
build.

If you are building this port with debug. Make sure you have enough memory and
space. :o Because when I did. I got a huge library and see the ld uses memory:

93282 root          1  -8    0   692M   680M biord    0:34 39.79% ld

You can test it by run 'make' or 'make install' without clean or remove work
directory, then go in work/WebKit*/Programs and run ./GtkLauncher. The
GtkLauncher is a mini browser for test, nothing fancy. I can add GtkLauncher
to install if anyone want me to. I can rename from GtkLauncher to
webkitgtk-mini-browser or something better.


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@10368 df743ca5-7f9a-e211-a948-0013205c9059