Gnote is a port of Tomboy to C++.
It is the same note taking application, including most of the add-ins (more
are to come). Synchronization support is being worked on.

WWW: http://live.gnome.org/Gnote

--

The reason why I am putting in here first is that... I haven't test it in the
runtime yet. Let me know if you have any problem with it. I will test it when
I get my desktop up and running again at some point. If it works great, it
will be replace from tomboy to gnote in our x11/gnome2. Reason:

	- The mono stuff sometime lag in FreeBSD, but not anymore with good
	  maintainers.
	- Tomboy is very limited on $ARCH (i386 and maybe amd64)
	- The gnote is heck a lot smaller to fit in CD and work on many $ARCH.


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@12496 df743ca5-7f9a-e211-a948-0013205c9059