Another day, another hal patch.
Attempt to workaround the GEOM lock race by sleeping half of a second when
a DESTROY event is seen by the storage subsystem.  Additionally, protect
against kern.geom.conftxt returning NULL.  When this happens, all disks
would get removed, and that's bad.

Fix a bug in the volume probing code which could lead to an assert crash.

Fix a bug in the volume probing code when it comes to handling UFS IDs.


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@13027 df743ca5-7f9a-e211-a948-0013205c9059