Update the pkg-message.
The adviced values are the double of what is default on 7.x and 8.x.
The default values where bumped in 9.x, and should be more then enough for
totem. So try to make clear that the pkg-message only goes for 7.x and 8.x.

adviced values in pkg-message:
kern.ipc.shmmax 67108864
kern.ipc.shmall 32768

What current has:
kern.ipc.shmmax 536870912
kern.ipc.shmall 131072


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@16106 df743ca5-7f9a-e211-a948-0013205c9059