Repocopy from farsight2. The newer empathy needs it.
The Farstream (formerly Farsight) project is an effort to create a framework
to deal with all known audio/video conferencing protocols. On one side it
offers a generic API that makes it possible to write plugins for different
streaming protocols, on the other side it offers an API for clients to use
those plugins.

WWW: http://www.freedesktop.org/wiki/Software/Farstream


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@16894 df743ca5-7f9a-e211-a948-0013205c9059