Remove ltverhack's hard depend on USE_AUTOTOOLS=libtool. If USE_AUTOTOOLS=libtool is defined it will restore the old behaviour.
The problem with the old situation was that it used libtool from LOCALBASE
and that libtool is configured to look in LOCALBASE first for linking.
This causes problems if the port you working on as API additions [1] which
don't exists in the installed version. The new method is to apply our
library version hack to the shipped ltmain.sh which will renerate a correct
libtool for us.

[1] This happend while trying to build the introspection files for the
Gstreamer 1.2 update, where libtool would link in the old installed libraries
and causing the build to fail with unresolved symbols. This problem has
been around for a long time but now I managed to track it now.
By actualy goofing up, and stumbeling into this :)


git-svn-id: svn://creme-brulee.marcuscom.com/ports/trunk@18886 df743ca5-7f9a-e211-a948-0013205c9059